Revamp maps
levine_towne_map: "Raw" map of Levine without any obstacles in the bump area.  Areas that do not have a high probability of walls are marked as free.  Areas with chairs are not blocked off.

levine_towne_nav: Same as levine_towne_map except there is a baseline occupancy for cells that are not free and certain areas are marked as occupied if chairs are present (e.g. bump area, some places in Towne).