Add linear ramp to motor driver
Performing the linear ramp in the node -- as oppposed to on the brainstem
-- requires using multiple threads: one to slowly change the setpoint when
a new velocity is commanded and one to continually read the current speed
and integrate the odometry.  Because both threads access the motor, access
to the AcronameMotor object and the motor state information are now
protected by a lock.